third_party/WebKit/LayoutTests/shadow-dom/v0/slotted-pseudo-element-in-v0v1-mixed.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="v0host">v0</div>
<div id="v1host"><span id="green">v1, should be green</span></div>
<script src='../resources/shadow-dom.js'></script>
<script>
// Test case for ensuring ::slotted() rules applied in v0/v1 mixed doc
// for catching any regression.
var v0root = v0host.createShadowRoot();
v0root.innerHTML = '<content></content>';

test(() => {
  var v1root = v1host.attachShadow({mode:'open'});
  v1root.innerHTML = '<style>::slotted(span) {color:green}</style><slot/>';
  var span = document.querySelector('#green');
  assert_equals(window.getComputedStyle(span).color, 'rgb(0, 128, 0)',
                'span element should be styled green.');
}, 'slotted rule should apply properly in v0/v1 mixed document.');
</script>
